requirements/dev.txt
# Everything the developer needs in addition to the production requirements
#-r prod.txt

# Testing
nose
py==1.4.20
Flask-Testing
factory-boy>=2.2.1
Flask-WebTest>=0.0.6

# Management script
Flask-Script

# Debug toolbar
Flask-DebugToolbar

flake8
wheel==0.22.0
coverage
